YaraScanDemo/requirement.txt
altgraph==0.17.2
atomicwrites==1.4.0
attrs==21.2.0
backports.entry-points-selectable==1.1.1
colorama==0.4.4
distlib==0.3.3
easygui==0.98.2
filelock==3.3.2
future==0.18.2
iniconfig==1.1.1
lxml==4.6.3
packaging==21.0
pefile==2021.9.3
platformdirs==2.4.0
pluggy==1.0.0
py==1.10.0
pyinstaller==4.6
pyinstaller-hooks-contrib==2021.3
pyparsing==2.4.7
PySide2==5.15.2
pytest==6.2.5
pywin32-ctypes==0.2.0
scandir==1.10.0
shiboken2==5.15.2
six==1.16.0
toml==0.10.2
virtualenv==20.10.0
yara-python==4.1.2
